backend/requirements.txt
Django==5.0
djangorestframework==3.15
psycopg2-binary==2.9.8
python-dotenv==1.0.0
djangorestframework-simplejwt==5.2.2
celery==5.3.1
redis==5.0.0
flake8==6.0.0